all5

+++ SRCS/draw_walls/draw_walls.c
@@ -43,6 +43,6 @@ int			draw_walls(t_vars *vars, int i, int last, float dist)
 	last = get_wall_pole(vars->wall_xy, vars->map, last);
 	init_drawconfig(vars, &drawconfig, last, vars->wall_xy);
 	draw_rect(vars, drawconfig, i,
-	vars->map->mapstruct.rate_height / (float)dist);
+	(float)(((float)vars->map->mapstruct.rate_height) / (float)dist));
 	return (last);
 }
